+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Uliweb Change Log
 
 * Add new app sequence, you can use it to create sequence value.
 * Fix generic version parameter passing to obj.save() bug, only when version is not False will be passed
+* Improve nginx and supervisor config content for deployment
 
 0.4 Version
 -----------------